requirements.txt
pyspark
fastapi
uvicorn[standard]
pydantic>=2
numpy
psycopg2-binary
pandas
pymongo
